update: fix trim logic and request server timeline

+++ lib/app/pages/media_suit/logic.dart
@@ -686,6 +686,7 @@ class MediaSuitController extends GetxController {
   //Multi Select
 
   var tempSelectedItems = <EditDataModel>[].obs;
+
   void confirmSelection() {
     for (var item in tempSelectedItems) {
       if (item.mediaClass == 1) {
@@ -700,7 +701,6 @@ class MediaSuitController extends GetxController {
     }
     tempSelectedItems.clear();
   }
-
   void addItemToTempList(String name, String url,widthVideoItem, String assetId, int mediaClass, {bool isloading = false}) {
     if (mediaClass == 1) {
       tempSelectedItems.add(EditDataModel(name, url, 0, assetId, 1 ,isloading: isloading));
@@ -709,7 +709,7 @@ class MediaSuitController extends GetxController {
     } else if (mediaClass == 3) {
       tempSelectedItems.add(EditDataModel(name, url, 0, assetId, 3 ,isloading: isloading));
     } else if (mediaClass == 4) {
-      tempSelectedItems.add(EditDataModel(name, url, widthVideoItem.toDouble(), assetId, 4, isloading: isloading));
+      tempSelectedItems.add(EditDataModel(name, url, widthVideoItem, assetId, 4, isloading: isloading));
 
     } else {
